move engine::func_builder::regmach into engine::regmach
This helps to dissect stack-machine and register-machine engine implementations.
