+++ crates/wasmi/src/engine/regmach/translator/result_mut.rs
