+++ crates/wasmi/src/engine/regmach/translator/stack/consts.rs
