+++ crates/wasmi/src/engine/regmach/translator/stack/provider.rs
