+++ crates/wasmi/src/engine/regmach/translator/utils.rs
